backend/requirements.txt
# FastAPI 核心
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.10.0
pydantic-settings>=2.6.0

# LangGraph 状态编排
langgraph>=0.2.0
langgraph-checkpoint>=2.0.0
langgraph-checkpoint-sqlite>=2.0.0
langchain-core>=0.3.0
langchain-openai>=0.3.0

# 数据库
sqlalchemy>=2.0.0
aiosqlite>=0.20.0

# HTTP 客户端
httpx>=0.27.0

# OpenAPI 解析
openapi-core>=0.19.0

# OpenTelemetry 遥测
opentelemetry-api>=1.27.0
opentelemetry-sdk>=1.27.0
opentelemetry-instrumentation-fastapi>=0.48.0
opentelemetry-instrumentation-httpx>=0.48.0
opentelemetry-exporter-otlp>=1.27.0

# 工具
python-dotenv>=1.0.0
python-multipart>=0.0.9

# 测试
pytest>=8.0.0
pytest-asyncio>=0.24.0
